1.master不kad安装离线包 2.离线包安装脚本cp判断文件是否存在

+++ rpminstall.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
-cp -r ./rpminstall.sh /opt/kad/down/rpms
+if [ ! -f "./rpminstall.sh" ];then
+	cp -r ./rpminstall.sh /opt/kad/down/rpms
+fi
 yum remove -y python3
 mode=${1:-3} 
 osname=(`uname -r`)
